Level Specific Crop Purchases

+++ Assets/Scripts/UI/StatusBar.cs
@@ -7,7 +7,7 @@ public class StatusBar : MonoBehaviour
 {
     [SerializeField] Slider bar;
 
-    public void Set(int curr, int max)
+    public void Set(float curr, int max)
     {
         bar.maxValue = max;
         bar.value = curr;
